# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Inventory adjustments dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /demand/summary:
    get:
      operationId: getDemandSummary
      tags: [demand]
      summary: Per-stock demand history, on-hand, lead time, and recommended min/max
      parameters:
        - name: monthsBack
          in: query
          schema: { type: integer, default: 6 }
        - name: serviceLevel
          in: query
          schema: { type: number, default: 0.95 }
        - name: demandCv
          in: query
          schema: { type: number }
        - name: leadTimeCv
          in: query
          schema: { type: number }
        - name: forecastWeeks
          in: query
          schema: { type: integer, default: 12 }
      responses:
        "200":
          description: Demand summary
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DemandSummary" }
  /demand/stock-detail:
    get:
      operationId: getDemandStockDetail
      tags: [demand]
      summary: Demand history + 12-week forecast for a single stock
      parameters:
        - name: stockId
          in: query
          required: true
          schema: { type: string }
        - name: monthsBack
          in: query
          schema: { type: integer, default: 6 }
        - name: bucket
          in: query
          schema: { $ref: "#/components/schemas/BucketSize" }
        - name: serviceLevel
          in: query
          schema: { type: number, default: 0.95 }
        - name: demandCv
          in: query
          schema: { type: number }
        - name: leadTimeCv
          in: query
          schema: { type: number }
        - name: forecastWeeks
          in: query
          schema: { type: integer, default: 12 }
      responses:
        "200":
          description: Stock demand detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DemandStockDetail" }
...
  /cycle-counts/complete:
    post:
      operationId: markCycleCountComplete
      tags: [cycle-counts]
      summary: Mark a stock as counted in a given week
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CycleCountCompleteInput" }
      responses:
        "200":
          description: Marked complete
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CycleCountCompletion" }
    delete:
      operationId: unmarkCycleCountComplete
      tags: [cycle-counts]
      summary: Remove the completion record for a stock + week
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CycleCountCompleteInput" }
      responses:
        "204":
          description: Removed
...
components:
...
  schemas:
...
    BucketSize:
      type: string
      enum: [day, week, month, quarter, year]
      default: week
...
    DemandStockMetrics:
      type: object
      required:
        - stockId
        - onHandFootage
        - onHandRollCount
        - totalDemandFootage
        - weeksOfHistory
        - avgWeeklyDemand
        - weeklyDemandStdDev
        - demandCv
        - autoDemandCv
        - demandCvOverridden
        - avgLeadTimeDays
        - autoLeadTimeDays
        - leadTimeDaysOverridden
        - leadTimeStdDev
        - leadTimeCv
        - autoLeadTimeCv
        - leadTimeCvOverridden
        - seasonalityWeights
        - defaultSeasonalityWeights
        - seasonalityWeightsOverridden
        - poObservations
        - typicalRollFootage
        - autoTypicalRollFootage
        - typicalRollFootageOverridden
        - safetyStockFootage
        - reorderPointFootage
        - maxFootage
        - suggestedOrderFootage
        - suggestedOrderRolls
        - belowMin
        - daysOfCover
        - forecast12wkFootage
        - openPoCount
        - openPoRolls
        - openPoFootage
        - activityStatus
        - customized
      properties:
        stockId: { type: string }
        description: { type: string, nullable: true }
        onHandFootage: { type: number }
        onHandRollCount: { type: integer }
        totalDemandFootage: { type: number }
        weeksOfHistory: { type: integer }
        avgWeeklyDemand: { type: number }
        weeklyDemandStdDev: { type: number }
        demandCv: { type: number, description: "Effective demand CV used in safety-stock math (override if set, else auto-derived)." }
        autoDemandCv: { type: number, description: "Auto-derived demand CV from history (stdev / mean of weekly demand)." }
        demandCvOverridden: { type: boolean, description: "True when demandCv comes from a per-stock override rather than the auto-derived value." }
        avgLeadTimeDays: { type: number, description: "Effective average lead time in days (override if set, else auto-derived from PO history)." }
        autoLeadTimeDays: { type: number, description: "Auto-derived average lead time in days from observed PO history." }
        leadTimeDaysOverridden: { type: boolean, description: "True when avgLeadTimeDays comes from a per-stock override rather than the auto-derived value." }
        leadTimeStdDev: { type: number }
        leadTimeCv: { type: number, description: "Effective lead-time CV used in safety-stock math (override if set, else auto-derived)." }
        autoLeadTimeCv: { type: number, description: "Auto-derived lead-time CV from observed PO lead times." }
        leadTimeCvOverridden: { type: boolean, description: "True when leadTimeCv comes from a per-stock override rather than the auto-derived value." }
        seasonalityWeights:
          type: array
          minItems: 3
          maxItems: 3
          items: { type: number }
          description: "Effective seasonality weights for months 1, 2, 3 of each quarter (sum to 1)."
        defaultSeasonalityWeights:
          type: array
          minItems: 3
          maxItems: 3
          items: { type: number }
          description: "Default seasonality weights used when no override is set."
        seasonalityWeightsOverridden: { type: boolean, description: "True when seasonalityWeights come from a per-stock override rather than the default." }
        poObservations: { type: integer }
        typicalRollFootage: { type: number, description: "Effective typical incoming roll size (override if set, else auto-derived)." }
        autoTypicalRollFootage: { type: number, description: "Auto-derived typical incoming roll size from observed PO history." }
        typicalRollFootageOverridden: { type: boolean, description: "True when typicalRollFootage comes from a per-stock override rather than the auto-derived value." }
        safetyStockFootage: { type: number }
        reorderPointFootage: { type: number }
        maxFootage: { type: number }
        suggestedOrderFootage: { type: number }
        suggestedOrderRolls: { type: integer }
        belowMin: { type: boolean }
        daysOfCover: { type: integer }
        forecast12wkFootage: { type: number }
        openPoCount: { type: integer, description: "Number of open (unreceived, non-cancelled) Stock-type purchase orders for this stock." }
        openPoRolls: { type: integer, description: "Sum of master-roll quantities across open POs." }
        openPoFootage: { type: number, description: "Estimated footage on order = openPoRolls * typicalRollFootage." }
        lastUsedDate: { type: string, format: date, nullable: true }
        daysSinceLastUse: { type: integer, nullable: true }
        activityStatus:
          type: string
          enum: [active, slowing, dormant, never]
          description: |
            Roll-up of recent demand activity:
            - active: used within the last 60 days
            - slowing: last used 60-180 days ago (review)
            - dormant: last used >180 days ago (consider inactivating)
            - never: no demand observed in window
        customized: { type: boolean, description: True if any per-stock forecast assumption override is set. }
    DemandSummary:
      type: object
      required: [windowFrom, windowTo, monthsBack, serviceLevel, forecastWeeks, generatedAt, items]
      properties:
        windowFrom: { type: string, format: date }
        windowTo: { type: string, format: date }
        monthsBack: { type: integer }
        serviceLevel: { type: number }
        forecastWeeks: { type: integer }
        generatedAt: { type: string, format: date-time }
        items:
          type: array
          items: { $ref: "#/components/schemas/DemandStockMetrics" }
    DemandPoint:
      type: object
      required: [periodStart, periodEnd, label, footage]
      properties:
        periodStart: { type: string, format: date }
        periodEnd: { type: string, format: date }
        label: { type: string }
        footage: { type: number }
    OpenPo:
      type: object
      required: [poNumber, stockId, quantityRolls]
      properties:
        poNumber: { type: string }
        stockId: { type: string }
        poDateIso: { type: string, format: date, nullable: true }
        quantityRolls: { type: integer, description: "Master-roll quantity ordered." }
        description: { type: string, nullable: true }
        daysOpen: { type: integer, nullable: true, description: "Days since the PO was placed." }
    DemandStockDetail:
      type: object
      required: [stockId, windowFrom, windowTo, bucket, serviceLevel, metrics, history, forecast, openPos]
      properties:
        stockId: { type: string }
        windowFrom: { type: string, format: date }
        windowTo: { type: string, format: date }
        bucket: { $ref: "#/components/schemas/BucketSize" }
        serviceLevel: { type: number }
        metrics: { $ref: "#/components/schemas/DemandStockMetrics" }
        history:
          type: array
          items: { $ref: "#/components/schemas/DemandPoint" }
        forecast:
          type: array
          items: { $ref: "#/components/schemas/DemandPoint" }
        openPos:
          type: array
          items: { $ref: "#/components/schemas/OpenPo" }
...
    CycleCountCompleteInput:
      type: object
      required: [stockId, quarter, week]
      properties:
        stockId: { type: string }
        quarter: { type: string }
        week: { type: integer }
        completedAt: { type: string, format: date-time, nullable: true }
    CycleCountCompletion:
      type: object
      required: [stockId, quarter, week, completedAt]
      properties:
        stockId: { type: string }
        quarter: { type: string }
        week: { type: integer }
        completedAt: { type: string, format: date-time }